Use uint64 (thanks carlclone and Harry Porter)

+++ kernel/proc.c
@@ -232,7 +232,7 @@ void userinit(void) {
 // Grow or shrink user memory by n bytes.
 // Return 0 on success, -1 on failure.
 int growproc(int n) {
-  uint sz;
+  uint64 sz;
   struct proc *p = myproc();
 
   sz = p->sz;
